add problem and testcase APIs with db-based judging

+++ codeforge/src/main/java/com/coding/codeforge/DTO/CodeRequest.java
@@ -4,9 +4,7 @@ public class CodeRequest {
 
     private String code;
     private String language;
-    private int problemId;
-
-    // GETTERS & SETTERS
+    private Long problemId;
 
     public String getCode() { return code; }
     public void setCode(String code) { this.code = code; }
@@ -14,6 +12,6 @@ public class CodeRequest {
     public String getLanguage() { return language; }
     public void setLanguage(String language) { this.language = language; }
 
-    public int getProblemId() { return problemId; }
-    public void setProblemId(int problemId) { this.problemId = problemId; }
+    public Long getProblemId() { return problemId; }
+    public void setProblemId(Long problemId) { this.problemId = problemId; }
 }
